session storage decorator
